+++ packages/dumb/src/components/tooltips/reservoir-tooltip/reservoir-tooltip.tsx
@@ -26,7 +26,6 @@
  */
 
 import React, { type FunctionComponent } from 'react'
-import i18next from 'i18next'
 import commonStyles from '../../../styles/tooltip-common.css'
 import {
   COMMON_TOOLTIP_TAIL_HEIGHT,
@@ -40,8 +39,9 @@ import {
 import { Tooltip } from '../../../index'
 import colors from '../../../styles/colors.css'
 import { getDateTitleForBaseDatum } from '../../../utils/tooltip/tooltip.util'
-import { PumpManufacturer, type ReservoirChange, type TimePrefs } from 'medical-domain'
+import { type ReservoirChange, type TimePrefs } from 'medical-domain'
 import { TooltipLine } from '../common/tooltip-line/tooltip-line'
+import { getReservoirChangeTitle } from '../../../utils/reservoir-change/reservoir-change.util'
 
 interface ReservoirTooltipProps {
   reservoir: ReservoirChange
@@ -50,35 +50,10 @@ interface ReservoirTooltipProps {
   timePrefs: TimePrefs
 }
 
-enum ChangeType {
-  Cartridge = 'cartridge',
-  Reservoir = 'reservoir'
-}
-
-const t = i18next.t.bind(i18next)
-
 export const ReservoirTooltip: FunctionComponent<ReservoirTooltipProps> = (props) => {
   const { reservoir, position, side, timePrefs } = props
 
-  const getChangeTypeByManufacturer = (manufacturer: PumpManufacturer): ChangeType => {
-    const manufacturerUpperCase = manufacturer.toUpperCase()
-    switch (manufacturerUpperCase) {
-      case PumpManufacturer.Sooil:
-        return ChangeType.Reservoir
-      case PumpManufacturer.Vicentra:
-      case PumpManufacturer.Roche:
-      case PumpManufacturer.Terumo:
-      case PumpManufacturer.Default:
-      default:
-        return ChangeType.Cartridge
-    }
-  }
-
-  const manufacturer = reservoir.pump?.manufacturer || PumpManufacturer.Default
-  const changeType: ChangeType = getChangeTypeByManufacturer(manufacturer)
-  const label = (changeType === ChangeType.Reservoir)
-    ? t('Reservoir change')
-    : t('Cartridge change')
+  const label = getReservoirChangeTitle(reservoir)
 
   return (
     <Tooltip