+++ packages/medical-domain/src/index.ts
@@ -98,6 +98,8 @@ import { RescueCarbsAveragePerRange } from './domains/models/statistics/carbs-st
 import { ManualBolusAveragePerRange } from './domains/models/statistics/basal-bolus-statistics.model'
 import { GROUP_ALARMS_THRESHOLD_MINUTES } from './domains/repositories/medical/datum/alarm-event.service'
 import { DeviceSystem } from './domains/models/medical/datum/enums/device-system.enum'
+import { DeviceEventSubtype } from './domains/models/medical/datum/enums/device-event-subtype.enum'
+import WarmUp from './domains/models/medical/datum/warm-up.model'
 
 export {
   AlarmCode,
@@ -125,6 +127,7 @@ export {
   type DateFilter,
   type Datum,
   type DeviceConfig,
+  DeviceEventSubtype,
   type DeviceParameterChange,
   DeviceSystem,
   type DurationValue,
@@ -167,6 +170,7 @@ export {
   CarbsStatisticsService,
   type ParametersChange,
   type PumpSettingsParameter,
+  WarmUp,
   ChangeType,
   BasalBolusStatisticsService,
   MS_IN_DAY,